Refactored Some code and changed location of one time use exceptions
