+++ src/main/java/ca/utoronto/fitbook/application/exceptions/UsernameAlreadyExistsException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 // The error thrown given a username already in the database
 @ResponseStatus(value= HttpStatus.CONFLICT, reason="Username already exists")
 public class UsernameAlreadyExistsException extends RuntimeException {
-    public UsernameAlreadyExistsException() {
-        super("Username already exists.");
+    public UsernameAlreadyExistsException(String username) {
+        super(String.format("Username %s already exists.", username));
     }
 }
