+++ src/test/java/ca/utoronto/fitbook/unit/UserRegisterTest.java
@@ -2,6 +2,7 @@ package ca.utoronto.fitbook.unit;
 
 import ca.utoronto.fitbook.BaseTest;
 import ca.utoronto.fitbook.adapter.persistence.localmemory.UserLocalMemoryRepository;
+import ca.utoronto.fitbook.application.exceptions.UsernameAlreadyExistsException;
 import ca.utoronto.fitbook.application.port.in.UserRegisterUseCase;
 import ca.utoronto.fitbook.application.port.in.command.UserProfileCommand;
 import ca.utoronto.fitbook.application.port.in.command.UserRegisterCommand;
@@ -33,35 +34,35 @@ public class UserRegisterTest extends BaseTest {
     @Test
     public void testPasswordNotMatch() {
         UserRegisterCommand jhon = new UserRegisterCommand("jhon", "12345678", "123456789");
-        Assertions.assertThrows(UserRegisterService.PasswordNotMatch.class, () -> userRegisterUseCase.createUser(jhon));
+        Assertions.assertThrows(UserRegisterService.PasswordNotMatchException.class, () -> userRegisterUseCase.createUser(jhon));
     }
 
     // Testing when the password is smaller than 8 characters
     @Test
     public void testPasswordTooShort() {
         UserRegisterCommand henny = new UserRegisterCommand("henny", "123", "123");
-        Assertions.assertThrows(UserRegisterService.PasswordTooShort.class, () -> userRegisterUseCase.createUser(henny));
+        Assertions.assertThrows(UserRegisterService.PasswordTooShortException.class, () -> userRegisterUseCase.createUser(henny));
     }
 
     // Testing when the password is bigger than 40 characters
     @Test
     public void testPasswordTooLong() {
         UserRegisterCommand cat = new UserRegisterCommand("cat", "12345678901234567890123456789012345678901", "12345678901234567890123456789012345678901");
-        Assertions.assertThrows(UserRegisterService.PasswordTooLong.class, () -> userRegisterUseCase.createUser(cat));
+        Assertions.assertThrows(UserRegisterService.PasswordTooLongException.class, () -> userRegisterUseCase.createUser(cat));
     }
 
     // Testing when the username is smaller than 3 characters
     @Test
     public void testUsernameTooShort() {
         UserRegisterCommand tom = new UserRegisterCommand("", "1234567", "1234567");
-        Assertions.assertThrows(UserRegisterService.NameTooShort.class, () -> userRegisterUseCase.createUser(tom));
+        Assertions.assertThrows(UserRegisterService.NameTooShortException.class, () -> userRegisterUseCase.createUser(tom));
     }
 
     // Testing when the username is bigger than 40 characters
     @Test
     public void testUsernameTooLong() {
         UserRegisterCommand bill = new UserRegisterCommand("BillBillBiBillBillBiBillBillBiBillBillBil", "123456789", "123456789");
-        Assertions.assertThrows(UserRegisterService.NameTooLong.class, () -> userRegisterUseCase.createUser(bill));
+        Assertions.assertThrows(UserRegisterService.NameTooLongException.class, () -> userRegisterUseCase.createUser(bill));
     }
 
     // Testing that a user is created if the username and passwords are correct
@@ -71,7 +72,6 @@ public class UserRegisterTest extends BaseTest {
         hanaId = userRegisterUseCase.createUser(hana).getId();
         Assertions.assertInstanceOf(String.class, hanaId);
         userLocalMemoryRepository.delete(hanaId);
-
     }
 
     // Testing when the username already exists
@@ -80,6 +80,6 @@ public class UserRegisterTest extends BaseTest {
         UserRegisterCommand tana = new UserRegisterCommand("tana", "123456789", "123456789");
         userRegisterUseCase.createUser(tana);
         UserRegisterCommand tanaAgain = new UserRegisterCommand("tana", "123123123", "123123123");
-        Assertions.assertThrows(UserRegisterService.UsernameAlreadyExists.class, () -> userRegisterUseCase.createUser(tanaAgain));
+        Assertions.assertThrows(UsernameAlreadyExistsException.class, () -> userRegisterUseCase.createUser(tanaAgain));
     }
 }
